requirements.txt
flask
twilio
flask-cors
python-dotenv
psycopg2